+++ app/core/middleware.php
@@ -1,0 +1,8 @@
+<?php
+
+return [
+    'global' => [
+        // List of middleware to be applied globally to all routes
+        \App\Middleware\RateLimitMiddleware::class => 10, // Allow up to 10 requests per minute
+    ]
+];